Add robot_query()
Executes SPARQL queries via ROBOT tool using R.

NOTE: rdflib package for R does not behave as expected for
SPARQL queries (not sure why) and cannot be used. While the
python rdflib import works and python functions can be sourced
into R, this is a messy approach that is likely to reduce
reproducibility. Instead, this function allows R to use ROBOT
in the system path (which most team members will have since it
is integral to development).