Save alliance disease file without datestamp
Instead, archives any existing file of the same name but
extracting the alliance version and file datetime and appending
that to the existing file. If an archived file of the same
version already exists, it checks that they are the same and
keeps the already archived version (if they differ, the process
errors).