Set rentrez version to github allenbaron/rentrez
This remote is currently a necessary dependency that ensures
rentrez::entrez_post() commands with >= 200 IDs are processed
correctly (via POST instead of GET) and reduce the need for ID
grouping/splitting/chunking.

Github PR #174 has been opened for ropensci/rentrez with this
fix. This remote dependency will be removed when that update
occurs.