Update args to handle NULLs/simplification [FAIL: 1]
UPDATE:
This approach results in a round trip conversion that results in
identical values, except in the case of NULL values, which are
converted to NA.

PLAN:
Could create a function that converts all NULL values to NA
first (which seems a more appropriate R placeholder for missing
values) and then this should work as hoped.

PROBLEM:
On further inspection, RJSONIO::fromJSON appears to be somewhat
unpredictable and difficult to debug. The nullvalue argument,
which should control what JSON null values are converted to when
read into R appears to do nothing. Instead, the simplify argument
when set to TRUE is responsible for the JSON null to R NA
conversion. There is no combinations of RJSONIO::Strict values
that work better than the current setup. To add to the problem,
RJSONIO is not under active development, so there's zero chance
this bug will be fixed (and it doesn't bode well for use in this
package, particularly given the goal of reproduciblity and
stability which these functions are to help achieve).

It might be worth searching jsonlite again for a solution.