Use jsonlite serialize method for list to json [PASS]
UPDATE:
jsonlite has the more verbose (un)serialize method for conversion
of R objects to JSON and it works perfectly for all tests without
any need for NULL/NA conversion. The only downside is it is fairly
verbose but I don't anticipate that being a problem since these
data will likely not be read directly by a person.