Modify pubmed_summary() to accept PMID list input
DESCRIPTION:
Modified to handle a list of PubMed IDs as input, along with
a vector of PMIDs or a web_history object.

PURPOSE:
This allows direct input from the results of a PubMed cited by
search with by_id = TRUE, which results in a named list of PMIDs
after extraction.