Fix naming of output list
PROBLEM:
Trying to ensure complete accuracy, the names were being extracted
from the API responses, with each response including all input
ids. However, if a response is empty because an article doesn't
have any listed citations yet and it is removed from the list,
the number of names would no longer match the number of results.

FIX:
Don't use the API response for names. Instead, name in the
function that calls the API using its input.