Create tests for format_subtree()
Tests pass, but results for fill_subclasses = FALSE are not as
expected.

PROBLEM:
Each existing duplicate should appear in the tree, only
the children of those duplicates should not; but no duplicates
are in final tree.

While this actually might be considered reasonable behavior,
it seems to vary with the makeup of the tree, which is going to
be very unpredicatable.

FIX RECOMMENDED:
Make fill_subclasses = TRUE obligatory (i.e. remove the parameter).