Create test for "no fill needed" error
PROBLEM:
format_subtree() internal function fill_subclass() fails at the
new column removal step when there is no duplication in the
original subtree because the new columns are never created (no
fill needed).

The original columns cannot be overwritten because the code in the
while loop depends on comparison with the previous step.