Revamp match_citations add_col
PROBLEM:
add_col was ignored when only 1 matching ID type was found in
data.frames but it should always work.

WEAKNESS:
Couldn't specify the name of the column to be added which meant,
I couldn't test multiple ID types separately and had to look up
the name to find it/avoid overwriting an existing column.

FIXES:
- add_col now always works when x is a data.frame
- if add_col is specified and x is NOT a df, errors early
- add_col input is now a string representing the column name to
be added