Allow for IRIs in read_term_file()
ROBOT term file allows for both CURIE and IRI formats. This update
brings this read function in line with that.

NOTE: There is no IRI or CURIE validation. In fact, the  regex
used to match is quite loose, extracting an 'identifier' from all
continuous non-space characters preceding a space and #.