Create inventory_omim() and deprecate onto_missing()
onto_missing() was designed to handle too many cases and was too
complex, only checking for missing OMIM mappings was ever
implemented. This is now handled by inventory_omim() with
improvements.

Output now is equivalent to omim_input with columns added that
specify whether the OMIM ID is present in the DO ('exists') and
columns with the relevant DO data (empty when it doesn't exist)..