Use polite full_request() in audit_url()
This works as intended with politely() but results can still be
lost because of errors in robotstxt internals (not sure what is
going on).

Using the `paths` argument of httr2::req_perform() is not
desirable. The cached files only contain the BODY of the response.

NEED to specify 'HEAD' request method!!!