Update is_valid_obo() & add is_valid_obo_prop()
**Potential breaking changes**

- Now requires namespaces to match OBO Foundry ontology prefix
(could cause breaking changes but only with misuse or if prefixes
in ROBOT are not being properly updated... may be delay).
- is_valid_obo() no longer recognizes properties (more likely to
be breaking change, but properties are rare)
- New is_valid_obo_prop() for validating OBO properties

DESIGN DECISION to split obo class & property validation because
most often classes are what are being tested for and it could be
a problem if properties slip through unnoticed. Can use both
predicates together if both are allowed.