Make glueV_cum() consistent with glueV
glue::glue() and glueV() identify expression string(s) as
unnamed inputs and temporary variables as named inputs.

This drops the requirement for explicitly stating the order of
expression strings using numbers as names for expression strings,
which is likely not intuitive, and makes inputs consistent with
other glue functions. It also adds the .sep argument.